+++ src/app/settings/projects/projects.component.ts
@@ -1,6 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Project } from '../../core/interfaces/project';
 import { ProjectsService } from '../../core/services/projects.service';
+import { MatDialog } from '@angular/material';
+import { DialogSynchronizeComponent } from '../../shared/dialog-synchronize/dialog-synchronize.component';
+import { DialogData } from '../../core/interfaces/dialog-data';
+import { DialogConfirmComponent } from '../../shared/dialog-confirm/dialog-confirm.component';
 
 @Component({
   selector: 'app-projects',
@@ -9,8 +13,9 @@ import { ProjectsService } from '../../core/services/projects.service';
 })
 export class ProjectsComponent implements OnInit {
   projects: Project[] = [];
+  dialogData: DialogData = {title: 'Project synchronization error'};
 
-  constructor(private projectsService: ProjectsService) {
+  constructor(private projectsService: ProjectsService, private dialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -20,12 +25,38 @@ export class ProjectsComponent implements OnInit {
   }
 
   delete(project: Project) {
-    this.projectsService.destroy(project).subscribe(() => {
-      this.projects.splice(this.projects.indexOf(project), 1);
+    const dialogRef = this.dialog.open(DialogConfirmComponent, {
+      width: '500px',
+      data: {title: 'Are you sure?', message: `Do you want do delete project ${project.Name}?`}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === true) {
+        this.projectsService.destroy(project).subscribe(() => {
+          this.projects.splice(this.projects.indexOf(project), 1);
+        });
+      }
+    });
+  }
+
+  openDialog(): void {
+    this.dialog.open(DialogSynchronizeComponent, {
+      width: '500px',
+      data: this.dialogData
     });
   }
 
   synchronize(project: Project) {
-    this.projectsService.synchronize(project).subscribe();
+    project.synchronizing = true;
+    this.projectsService.synchronize(project).subscribe(() => {
+      project.synchronizationStatus = 'success';
+      project.synchronizing = false;
+    }, (error) => {
+      this.dialogData.message = error.error;
+      project.synchronizationStatus = 'error';
+      project.synchronizing = false;
+      this.openDialog();
+    }, () => {
+      project.synchronizing = false;
+    });
   }
 }
