+++ src/app/settings/repositories/repositories.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Repository } from '../../core/interfaces/repository';
 import { RepositoriesService } from '../../core/services/repositories.service';
+import { DialogConfirmComponent } from '../../shared/dialog-confirm/dialog-confirm.component';
+import { MatDialog } from '@angular/material';
 
 @Component({
   selector: 'app-repositories',
@@ -10,7 +12,7 @@ import { RepositoriesService } from '../../core/services/repositories.service';
 export class RepositoriesComponent implements OnInit {
   repositories: Repository[] = [];
 
-  constructor(private repositoriesService: RepositoriesService) {
+  constructor(private repositoriesService: RepositoriesService, private dialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -20,8 +22,16 @@ export class RepositoriesComponent implements OnInit {
   }
 
   delete(repository: Repository) {
-    this.repositoriesService.destroy(repository).subscribe(() => {
-      this.repositories.splice(this.repositories.indexOf(repository), 1);
+    const dialogRef = this.dialog.open(DialogConfirmComponent, {
+      width: '500px',
+      data: {title: 'Are you sure?', message: `Do you want do delete repository ${repository.Name}?`}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === true) {
+        this.repositoriesService.destroy(repository).subscribe(() => {
+          this.repositories.splice(this.repositories.indexOf(repository), 1);
+        });
+      }
     });
   }
 }
